Add fixes to init frag_size and clear send_msgoff in req struct
Reason for fixing frag_size initialization-
tidflows that are only init-ed for ACKs and NAKs were not being
initialized with frag_size, which is checked in an assert in
ips_proto_send_ctrl_message(). This caused the assert to fail.
We now initialize frag_size in these tidflows to SDMA MTU.

Reason for clearing send_msgoff-
The criteria for completing an isend() was for send_msgoff
to be larger or euqal to send_msglen. This worked fine for
small workloads where psm2_mq_reqs were not recycled often
enough. However, if the workload ran the same OPCODE_SHORT
path through isend() repeatedly, send_msgoff would eventually
overflow and then the criterion for isend() would fail.
Fixing that here and also removing some dead code.

Signed-off-by: Aravind Gopalakrishnan <Aravind.Gopalakrishnan@intel.com>